remove config source and update rust repo
