+++ crates/settings/src/lib.rs
@@ -2,14 +2,11 @@ pub mod accounts;
 pub mod blocks;
 pub mod chart;
 pub mod config;
-pub mod config_source;
 pub mod deployer;
 pub mod deployment;
 pub mod gui;
-pub mod merge;
 pub mod metaboard;
 pub mod network;
-pub mod new_config;
 pub mod order;
 pub mod orderbook;
 pub mod plot_source;
@@ -25,7 +22,6 @@ pub mod unit_test; // NOTE: this is used by unit_tests module in common crate
 pub mod yaml;
 
 pub(crate) use chart::*;
-pub(crate) use config_source::*;
 pub(crate) use deployer::*;
 pub(crate) use deployment::*;
 pub(crate) use gui::*;
