+++ crates/settings/src/scenario.rs
@@ -38,16 +38,16 @@ pub struct ScenarioCfg {
 impl_wasm_traits!(ScenarioCfg);
 
 impl ScenarioCfg {
-    pub fn validate_runs(value: &str) -> Result<u64, ParseScenarioConfigSourceError> {
+    pub fn validate_runs(value: &str) -> Result<u64, ParseScenarioConfigError> {
         value
             .parse::<u64>()
-            .map_err(ParseScenarioConfigSourceError::RunsParseError)
+            .map_err(ParseScenarioConfigError::RunsParseError)
     }
 
-    pub fn validate_blocks(value: &str) -> Result<BlocksCfg, ParseScenarioConfigSourceError> {
+    pub fn validate_blocks(value: &str) -> Result<BlocksCfg, ParseScenarioConfigError> {
         match serde_yaml::from_str::<BlocksCfg>(value) {
             Ok(blocks) => Ok(blocks),
-            Err(_) => Err(ParseScenarioConfigSourceError::BlocksParseError(
+            Err(_) => Err(ParseScenarioConfigError::BlocksParseError(
                 value.to_string(),
             )),
         }
@@ -94,8 +94,8 @@ impl ScenarioCfg {
             if let Some(parent_value) = parent_scenario.bindings.as_ref().and_then(|pb| pb.get(&k))
             {
                 if *parent_value != v {
-                    return Err(YamlError::ParseScenarioConfigSourceError(
-                        ParseScenarioConfigSourceError::ParentBindingShadowedError(k.to_string()),
+                    return Err(YamlError::ParseScenarioConfigError(
+                        ParseScenarioConfigError::ParentBindingShadowedError(k.to_string()),
                     ));
                 }
             }
@@ -155,8 +155,8 @@ impl ScenarioCfg {
         if let Some(current_deployer) = current_deployer {
             if let Some(parent_deployer) = parent_scenario.deployer.as_ref() {
                 if current_deployer.key != parent_deployer.key {
-                    return Err(YamlError::ParseScenarioConfigSourceError(
-                        ParseScenarioConfigSourceError::ParentDeployerShadowedError(
+                    return Err(YamlError::ParseScenarioConfigError(
+                        ParseScenarioConfigError::ParentDeployerShadowedError(
                             current_deployer.key.clone(),
                         ),
                     ));
@@ -184,9 +184,9 @@ impl ScenarioCfg {
                 bindings: bindings.clone(),
                 runs,
                 blocks,
-                deployer: deployer.clone().ok_or(
-                    ParseScenarioConfigSourceError::DeployerNotFound(scenario_key),
-                )?,
+                deployer: deployer
+                    .clone()
+                    .ok_or(ParseScenarioConfigError::DeployerNotFound(scenario_key))?,
             },
         );
 
@@ -383,8 +383,8 @@ impl YamlParsableHash for ScenarioCfg {
                     )?;
 
                     if deployer.is_none() {
-                        return Err(YamlError::ParseScenarioConfigSourceError(
-                            ParseScenarioConfigSourceError::DeployerNotFound(scenario_key),
+                        return Err(YamlError::ParseScenarioConfigError(
+                            ParseScenarioConfigError::DeployerNotFound(scenario_key),
                         ));
                     }
                 }
@@ -426,7 +426,7 @@ impl PartialEq for ScenarioCfg {
 }
 
 #[derive(Error, Debug, PartialEq)]
-pub enum ParseScenarioConfigSourceError {
+pub enum ParseScenarioConfigError {
     #[error("Failed to parse runs")]
     RunsParseError(ParseIntError),
     #[error("Parent binding shadowed by child: {0}")]
@@ -441,20 +441,20 @@ pub enum ParseScenarioConfigSourceError {
     BlocksParseError(String),
 }
 
-impl ParseScenarioConfigSourceError {
+impl ParseScenarioConfigError {
     pub fn to_readable_msg(&self) -> String {
         match self {
-            ParseScenarioConfigSourceError::RunsParseError(err) =>
+            ParseScenarioConfigError::RunsParseError(err) =>
                 format!("The 'runs' value in your scenario YAML configuration must be a valid number: {}", err),
-            ParseScenarioConfigSourceError::ParentBindingShadowedError(binding) =>
+            ParseScenarioConfigError::ParentBindingShadowedError(binding) =>
                 format!("Binding conflict in your YAML configuration: The child scenario is trying to override the binding '{}' that was already defined in a parent scenario. Child scenarios cannot change binding values defined by parents.", binding),
-            ParseScenarioConfigSourceError::ParentDeployerShadowedError(deployer) =>
+            ParseScenarioConfigError::ParentDeployerShadowedError(deployer) =>
                 format!("Deployer conflict in your YAML configuration: The child scenario is trying to use deployer '{}' which differs from the deployer specified in the parent scenario. Child scenarios must use the same deployer as their parent.", deployer),
-            ParseScenarioConfigSourceError::DeployerNotFound(scenario) =>
+            ParseScenarioConfigError::DeployerNotFound(scenario) =>
                 format!("No deployer was found for scenario '{}' in your YAML configuration. Please specify a deployer for this scenario or ensure it inherits one from a parent scenario.", scenario),
-            ParseScenarioConfigSourceError::ParentOrderbookShadowedError(orderbook) =>
+            ParseScenarioConfigError::ParentOrderbookShadowedError(orderbook) =>
                 format!("Orderbook conflict in your YAML configuration: The child scenario is trying to use orderbook '{}' which differs from the orderbook specified in the parent scenario. Child scenarios must use the same orderbook as their parent.", orderbook),
-            ParseScenarioConfigSourceError::BlocksParseError(blocks) =>
+            ParseScenarioConfigError::BlocksParseError(blocks) =>
                 format!("Failed to parse the 'blocks' configuration in your YAML: {}. Please ensure it follows the correct format.", blocks),
         }
     }
@@ -466,258 +466,11 @@ pub struct ScenarioParent {
     bindings: Option<HashMap<String, String>>,
     deployer: Option<Arc<DeployerCfg>>,
 }
-
-// Shadowing is disallowed for deployers, orderbooks and specific bindings.
-// If a child specifies one that is already set by the parent, this is an error.
-//
-// Nested scenarios within the ScenarioConfigSource struct are flattened out into a
-// hashmap of scenarios, where the key is the path such as foo.bar.baz.
-// Every level of the scenario path inherits its parents bindings recursively.
-impl ScenarioConfigSource {
-    pub fn try_into_scenarios(
-        &self,
-        name: String,
-        parent: &ScenarioParent,
-        deployers: &HashMap<String, Arc<DeployerCfg>>,
-    ) -> Result<HashMap<String, Arc<ScenarioCfg>>, ParseScenarioConfigSourceError> {
-        // Determine the resolved name for the deployer, preferring the explicit deployer name if provided.
-        let resolved_name = self.deployer.as_ref().unwrap_or(&name);
-
-        // Attempt to find the deployer using the resolved name.
-        let resolved_deployer = deployers.get(resolved_name);
-
-        // If no deployer is found using the resolved name, fall back to the parent's deployer, if any.
-        let deployer_ref = resolved_deployer.or(parent.deployer.as_ref());
-
-        // If no deployer could be resolved and there's no parent deployer, return an error.
-        let deployer_ref = deployer_ref
-            .ok_or_else(|| ParseScenarioConfigSourceError::DeployerNotFound(name.clone()))?;
-
-        // Check for non-matching override: if both the current and parent deployers are present and different, it's an error.
-        if let (deployer, Some(parent_deployer)) = (deployer_ref, parent.deployer.as_ref()) {
-            if deployer.key != parent_deployer.key {
-                return Err(ParseScenarioConfigSourceError::ParentDeployerShadowedError(
-                    resolved_name.clone(),
-                ));
-            }
-        }
-
-        // Merge bindings and check for shadowing
-        let mut bindings = parent
-            .bindings
-            .as_ref()
-            .map_or_else(HashMap::new, |pb| pb.clone());
-        for (k, v) in &self.bindings {
-            if let Some(parent_value) = parent.bindings.as_ref().and_then(|pb| pb.get(k)) {
-                if parent_value != v {
-                    return Err(ParseScenarioConfigSourceError::ParentBindingShadowedError(
-                        k.to_string(),
-                    ));
-                }
-            }
-            bindings.insert(k.to_string(), v.to_string());
-        }
-
-        // Create and add the parent scenario for this level
-        let parent_scenario = Arc::new(ScenarioCfg {
-            document: Arc::new(RwLock::new(StrictYaml::String("".to_string()))),
-            key: name.clone(),
-            bindings: bindings.clone(),
-            runs: self.runs,
-            blocks: self.blocks.clone(),
-            deployer: deployer_ref.clone(),
-        });
-
-        let mut scenarios = HashMap::new();
-        scenarios.insert(name.clone(), parent_scenario);
-
-        // Recursively add child scenarios
-        if let Some(scenarios_map) = &self.scenarios {
-            for (child_name, child_scenario) in scenarios_map {
-                let child_scenarios = child_scenario.try_into_scenarios(
-                    format!("{}.{}", name, child_name),
-                    &ScenarioParent {
-                        key: "".to_string(),
-                        bindings: Some(bindings.clone()),
-                        deployer: Some(deployer_ref.clone()),
-                    },
-                    deployers,
-                )?;
-
-                scenarios.extend(child_scenarios);
-            }
-        }
-
-        Ok(scenarios)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
-        blocks::{BlockCfg, BlockRangeCfg},
-        spec_version::SpecVersion,
-        test::mock_deployer,
-    };
-    use alloy::primitives::Address;
-    use std::collections::HashMap;
-    use url::Url;
+    use crate::blocks::{BlockCfg, BlockRangeCfg};
     use yaml::tests::get_document;
-
-    #[test]
-    fn test_scenarios_conversion_with_nesting() {
-        // Initialize networks as in the previous example
-        let mut networks = HashMap::new();
-        networks.insert(
-            "mainnet".to_string(),
-            NetworkConfigSource {
-                rpc: Url::parse("https://mainnet.node").unwrap(),
-                chain_id: 1,
-                label: Some("Ethereum Mainnet".to_string()),
-                network_id: Some(1),
-                currency: Some("ETH".to_string()),
-            },
-        );
-
-        // Define a deployer
-        let mut deployers = HashMap::new();
-        deployers.insert(
-            "mainnet".to_string(),
-            DeployerConfigSource {
-                address: "0xabcdef0123456789ABCDEF0123456789ABCDEF01"
-                    .parse::<Address>()
-                    .unwrap(),
-                network: None,
-                label: Some("Mainnet Deployer".to_string()),
-            },
-        );
-
-        // Define nested scenarios
-        let mut nested_scenario2 = HashMap::new();
-        nested_scenario2.insert(
-            "nested_scenario2".to_string(),
-            ScenarioConfigSource {
-                bindings: HashMap::new(), // Assuming no bindings for simplification
-                runs: Some(2),
-                blocks: None,
-                deployer: None,
-                scenarios: None, // No further nesting
-            },
-        );
-
-        let mut nested_scenario1 = HashMap::new();
-        nested_scenario1.insert(
-            "nested_scenario1".to_string(),
-            ScenarioConfigSource {
-                bindings: HashMap::new(), // Assuming no bindings for simplification
-                runs: Some(5),
-                blocks: None,
-                deployer: None,
-                scenarios: Some(nested_scenario2), // Include nested_scenario2
-            },
-        );
-
-        // Define root scenario with nested_scenario1
-        let mut scenarios = HashMap::new();
-        scenarios.insert(
-            "root_scenario".to_string(),
-            ScenarioConfigSource {
-                bindings: HashMap::new(), // Assuming no bindings for simplification
-                runs: Some(10),
-                blocks: None,
-                deployer: Some("mainnet".to_string()),
-                scenarios: Some(nested_scenario1), // Include nested_scenario1
-            },
-        );
-
-        // Construct ConfigSource with the above scenarios
-        let config_string = ConfigSource {
-            version: SpecVersion::current().to_string(),
-            using_networks_from: HashMap::new(),
-            networks,
-            subgraphs: HashMap::new(), // Assuming no subgraphs for simplification
-            metaboards: HashMap::new(), // Assuming no metaboards for simplification
-            orderbooks: HashMap::new(), // Assuming no orderbooks for simplification
-            tokens: HashMap::new(),    // Assuming no tokens for simplification
-            deployers,
-            orders: HashMap::new(), // Assuming no orders for simplification
-            scenarios,
-            charts: HashMap::new(), // Assuming no charts for simplification
-            deployments: HashMap::new(),
-            sentry: None,
-            accounts: None, // Assuming no accounts for simplification
-            gui: None,
-        };
-
-        // Perform the conversion
-        let config_result = Config::try_from(config_string);
-        assert!(config_result.is_ok());
-
-        let config = config_result.unwrap();
-
-        // Verify the root scenario
-        assert!(config.scenarios.contains_key("root_scenario"));
-        let root_scenario = config.scenarios.get("root_scenario").unwrap();
-        assert_eq!(root_scenario.runs, Some(10));
-
-        // Verify the first level of nested scenarios
-        assert!(config
-            .scenarios
-            .contains_key("root_scenario.nested_scenario1"));
-        let nested_scenario1 = config
-            .scenarios
-            .get("root_scenario.nested_scenario1")
-            .unwrap();
-        assert_eq!(nested_scenario1.runs, Some(5));
-
-        // Verify the second level of nested scenarios
-        assert!(config
-            .scenarios
-            .contains_key("root_scenario.nested_scenario1.nested_scenario2"));
-        let nested_scenario2 = config
-            .scenarios
-            .get("root_scenario.nested_scenario1.nested_scenario2")
-            .unwrap();
-        assert_eq!(nested_scenario2.runs, Some(2));
-    }
-
-    #[test]
-    fn test_scenario_shadowing_error_in_bindings() {
-        let parent_bindings =
-            HashMap::from([("shared_key".to_string(), "parent_value".to_string())]);
-
-        let parent_scenario = ScenarioParent {
-            key: "".to_string(),
-            bindings: Some(parent_bindings),
-            deployer: Some(mock_deployer()),
-        };
-
-        let mut child_bindings = HashMap::new();
-        child_bindings.insert("shared_key".to_string(), "child_value".to_string()); // Intentionally shadowing parent binding
-
-        let child_scenario = ScenarioConfigSource {
-            bindings: child_bindings,
-            runs: None,
-            blocks: None,
-            deployer: None,
-            scenarios: None,
-        };
-
-        let result = child_scenario.try_into_scenarios(
-            "child".to_string(),
-            &parent_scenario,
-            &HashMap::new(), // Empty deployers for simplification
-        );
-
-        assert!(result.is_err());
-        match result.err().unwrap() {
-            ParseScenarioConfigSourceError::ParentBindingShadowedError(key) => {
-                assert_eq!(key, "shared_key");
-            }
-            _ => panic!("Expected ParentBindingShadowedError"),
-        }
-    }
 
     #[test]
     fn test_parse_scenarios_from_yaml() {
@@ -829,8 +582,8 @@ scenarios:
         let error = ScenarioCfg::parse_all_from_yaml(vec![get_document(yaml)], None).unwrap_err();
         assert_eq!(
             error.to_string(),
-            YamlError::ParseScenarioConfigSourceError(
-                ParseScenarioConfigSourceError::ParentBindingShadowedError("key1".to_string())
+            YamlError::ParseScenarioConfigError(
+                ParseScenarioConfigError::ParentBindingShadowedError("key1".to_string())
             )
             .to_string()
         );
@@ -865,8 +618,8 @@ scenarios:
         let error = ScenarioCfg::parse_all_from_yaml(vec![get_document(yaml)], None).unwrap_err();
         assert_eq!(
             error.to_string(),
-            YamlError::ParseScenarioConfigSourceError(
-                ParseScenarioConfigSourceError::ParentDeployerShadowedError("testnet".to_string())
+            YamlError::ParseScenarioConfigError(
+                ParseScenarioConfigError::ParentDeployerShadowedError("testnet".to_string())
             )
             .to_string()
         );
