+++ tauri-app/src-tauri/src/commands/mod.rs
@@ -2,7 +2,6 @@ pub mod app;
 pub mod authoring_meta;
 pub mod chain;
 pub mod charts;
-pub mod config;
 pub mod dotrain;
 pub mod dotrain_add_order_lsp;
 pub mod order;
