+++ tauri-app/src-tauri/src/error.rs
@@ -2,10 +2,7 @@ use alloy::hex::FromHexError;
 use alloy::primitives::ruint::{FromUintError, ParseError as FromUintParseError};
 use alloy_ethers_typecast::{client::LedgerClientError, transaction::ReadableClientError};
 use dotrain::error::ComposeError;
-use rain_orderbook_app_settings::config::ParseConfigSourceError;
-use rain_orderbook_app_settings::config_source::ConfigSourceError;
-use rain_orderbook_app_settings::merge::MergeError;
-use rain_orderbook_app_settings::new_config::ParseConfigError;
+use rain_orderbook_app_settings::ParseConfigError;
 use rain_orderbook_common::dotrain_order::DotrainOrderError;
 use rain_orderbook_common::fuzz::FuzzRunnerError;
 use rain_orderbook_common::remove_order::RemoveOrderArgsError;
@@ -59,12 +56,6 @@ pub enum CommandError {
     FuzzRunnerError(#[from] FuzzRunnerError),
 
     #[error(transparent)]
-    MergeError(#[from] MergeError),
-
-    #[error(transparent)]
-    ParseConfigSourceError(#[from] ParseConfigSourceError),
-
-    #[error(transparent)]
     ParseConfigYamlError(#[from] serde_yaml::Error),
 
     #[error(transparent)]
@@ -75,9 +66,6 @@ pub enum CommandError {
 
     #[error(transparent)]
     ComposeError(#[from] ComposeError),
-
-    #[error(transparent)]
-    ConfigSourceError(#[from] ConfigSourceError),
 
     #[error(transparent)]
     DotrainOrderError(#[from] DotrainOrderError),
@@ -112,7 +100,7 @@ pub enum CommandError {
     RainEvalResultError(#[from] rain_orderbook_common::fuzz::RainEvalResultError),
 
     #[error(transparent)]
-    NewConfigError(#[from] ParseConfigError),
+    ParseConfigError(#[from] ParseConfigError),
 }
 
 impl Serialize for CommandError {
